+++ Source/Models/TransactionRequest.swift
@@ -6,6 +6,8 @@
 //  Copyright © 2017-2018 Omise Go Pte. Ltd. All rights reserved.
 //
 
+import BigInt
+
 /// The different types of request that can be generated
 ///
 /// - receive: The requester wants to receive an amount of token
@@ -37,7 +39,7 @@ public struct TransactionRequest {
     public let token: Token
     /// The amount of token to use for the transaction (down to subunit to unit)
     /// This amount needs to be either specified by the requester or the consumer
-    public let amount: Double?
+    public let amount: BigInt?
     /// The address from which to send or receive the tokens
     public let address: String
     /// The user that initiated the request
@@ -113,7 +115,7 @@ extension TransactionRequest: Decodable {
         id = try container.decode(String.self, forKey: .id)
         type = try container.decode(TransactionRequestType.self, forKey: .type)
         token = try container.decode(Token.self, forKey: .token)
-        amount = try container.decodeIfPresent(Double.self, forKey: .amount)
+        amount = try container.decodeIfPresent(BigInt.self, forKey: .amount)
         address = try container.decode(String.self, forKey: .address)
         user = try container.decodeIfPresent(User.self, forKey: .user)
         account = try container.decodeIfPresent(Account.self, forKey: .account)
